Move all WIP into test folder
