+++ src/_routes/StandaloneRoutes.js
@@ -27,7 +27,7 @@ import DigitStatsCard from '../digit-stats/DigitStatsCard';
 import { requireAuthOnEnter } from '../_data/Auth';
 
 import StatementTesting from '../contract-details/ForTesting';
-import Single from '../trade/SingleTradeXCard';
+import Single from '../trade/test/SingleTradeXCard';
 
 export default [
     { path: 'trades', component: TradesPage, onEnter: requireAuthOnEnter },
