+++ src/trade/test/SingleTradeXCard.js
