+++ src/trade/test/TradeXCardContainer.js
